shader + renderer + structural changes

+++ include/camera/camera.hpp
@@ -26,7 +26,7 @@ public:
     virtual ~camera_t();
 
     virtual std::optional<ray::ray_t> generate_ray(
-        int const x, int const y,
+        size_t const x, size_t const y,
         float const* const cam_jitter = nullptr
     ) const = 0;
 };
